// client/src/app/shared/Models/order.ts
export interface ICreateOrder { 
    userName: string; 
    totalPrice: number; 
    firstName: string;
     lastName: string;
      emailAddress: string;
       addressLine: string; 
       country: string; 
       state: string; 
       zipCode: string; 
       cardName: string;
        cardNumber: string; 
        expiration: string; 
        cvv: string; 
        paymentMethod: number; }